Move "path is optional" logic into FindPythonDefinition.
This will allow us to expose it to more commands.

The prompt I used (and only had to do minor post-AI edits):

Change FindPythonDefinition (src/select_python.py) so that the path is
Optional[str]. If the path is not provided, the semantics should be to search in
all the files (allowed by file_access_policy).

See the implementation in ReplacePythonCommand that does this (in
src/replace_python_command.py). Just move that logic from ReplacePythonCommand
into FindPythonDefinition.

ReplacePythonCommand should simply do:

  matches : List[Selection] = FindPythonDefinition(..., path, identifier)

Instead of duplicating this logic that we're moving.