list_files now includes more details, maybe they are useful.
I was incredibly disappointed with the AI to implement this change.
It cheated in many ways.
In the end I gave up and decided to implement this myself.

It's interesting that it generates such strong negative emotions (of "OMG how
can you be so incompetent!") in me. I guess because I thought this would be a
trivial change.

Prompt used:

I want to change ListFilesCommand to not only list files, but also include a bit of information about them. For now, I want the lines to look like this:

  src/foo.py: 512 lines, 4939 bytes

Look at src/list_files_command.py and extend the code there to generate this information.

Add a unit test (should be in src/test_list_files_command.py) that validates this. Perhaps it should create a test directory in /tmp with two example files.